refactor: move the hidden textarea in the ultron-web
